add slider script to display/hide flowers

+++ flower_app/models.py
@@ -3,8 +3,8 @@ from django.db import models
 
 class Flower(models.Model):
     name = models.CharField(max_length=100, unique=True)
-    x_position = models.FloatField()
-    y_position = models.FloatField()
+    x_position = models.IntegerField()
+    y_position = models.IntegerField()
     start_date = models.PositiveIntegerField()
     end_date = models.PositiveIntegerField()
     image_url = models.URLField(max_length=200)  # Change to URLField
